MAINT: Resolve small changes for frame module refactoring.
