+++ celest/coordinates/transforms.py
@@ -1,10 +1,10 @@
 
 
 from celest.coordinates.astronomical_quantities import earth_rotation_angle
-from celest.coordinates.azel import AzEl
-from celest.coordinates.gcrs import GCRS
-from celest.coordinates.itrs import ITRS
-from celest.coordinates.wgs84 import WGS84
+from celest.coordinates.frames.azel import AzEl
+from celest.coordinates.frames.gcrs import GCRS
+from celest.coordinates.frames.itrs import ITRS
+from celest.coordinates.frames.wgs84 import WGS84
 from celest.coordinates.nutation_precession_matrices import (
     bias_matrix,
     precession_matrix,
